+++ src/assets/locales/en/wp.js
@@ -1,0 +1,14 @@
+export default {
+    actions: {
+        create: 'Create new Client',
+    },
+    fields: {
+        id: 'Phone',
+        alias: 'Alias'
+    },
+    placeholders: {
+        delete: 'The client will be removed, do ypu want to continue?',
+        default: 'Default client to confirmations',
+        select_default: 'Select default'
+    }
+}