+++ src/assets/locales/es/wp.js
@@ -1,0 +1,14 @@
+export default {
+    actions: {
+        create: 'Crear nuevo cliente',
+    },
+    fields: {
+        id: 'Teléfono',
+        alias: 'Alias'
+    },
+    placeholders: {
+        delete: 'El cliente se eliminará, ¿desea continuar?',
+        default: 'Cliente por defecto para las confirmaciones',
+        select_default: 'Selecccionar como cliente por defecto'
+    }
+}