+++ src/types/ServiceInterface.ts
@@ -13,6 +13,7 @@ export interface ServiceInterface {
   comment: string | null
   amount: number | null
   metadata: Metadata | null
+  wp_client_id: string
   driver_id: string | null
   client_id: string | null
   created_at: number