+++ src/types/WpClient.ts
@@ -1,0 +1,6 @@
+export type WpClient = {
+  id: string
+  alias: string
+  wpNotifications: boolean
+  chatBot: boolean
+}