+++ tests/testSetup.ts
@@ -6,6 +6,7 @@ import {Server} from 'socket.io'
 import WhatsAppClient from '@/services/gordaApi/WhatsAppClient'
 import {config, enableAutoUnmount} from '@vue/test-utils'
 import {createPinia, setActivePinia} from 'pinia'
+import {WpClient} from '@/types/WpClient'
 
 require('./mocks/maps/googleMaps')
 
@@ -53,6 +54,7 @@ jest.mock('firebase/database', () => {
 		}),
 		off: jest.fn(),
 		set: jest.fn(),
+    remove: jest.fn(),
     child: jest.fn()
   }
 })
@@ -72,6 +74,11 @@ jest.mock('firebase/firestore', () => {
     where: jest.fn(),
     orderBy: jest.fn(),
     limit: jest.fn(),
+    limitToLast: jest.fn(),
+    startAt: jest.fn(),
+    startAfter: jest.fn(),
+    endBefore: jest.fn(),
+    getCountFromServer: jest.fn(),
     onSnapshot: jest.fn().mockReturnValue({
       forEach: (callback: any): void => { callback() }
     }),
@@ -80,6 +87,7 @@ jest.mock('firebase/firestore', () => {
     deleteDoc: jest.fn(),
   };
 });
+
 jest.mock('firebase/storage')
 jest.mock('qrcode')
 
@@ -106,12 +114,17 @@ let server: httpServer
 function openServer(done: jest.DoneCallback): void {
   server = createServer()
   socket = new Server(server)
-  server.listen(process.env.VUE_APP_WP_CLIENT_API_PORT ?? 3000,() => {
+  server.listen(3000,() => {
     socket.on('connection', () => {
       done()
     })
   })
-  WhatsAppClient.getInstance()
+  WhatsAppClient.getInstance({
+    id: '3103794656',
+    wpNotifications: false,
+    chatBot: false,
+    alias: 'Test'
+  } as WpClient)
 }
 
 export {