+++ tests/unit/components/services/CreateService.spec.ts
@@ -14,6 +14,7 @@ import {getPlaces} from '../../../mocks/entities/PlaceMock'
 import {usePlacesStore} from '@/services/stores/PlacesStore'
 import {useClientsStore} from '@/services/stores/ClientsStore'
 import {StrHelper} from '@/helpers/StrHelper'
+import {useWpClientsStore} from "@/services/stores/WpClientStore";
 
 describe('CreateService.vue', () => {
 	let wrapper: VueWrapper<any>
@@ -22,6 +23,7 @@ describe('CreateService.vue', () => {
 	beforeEach(async () => {
 		const placesStore = usePlacesStore()
 		const clientsStore = useClientsStore()
+		const wpClient = useWpClientsStore()
 		wrapper = mount(CreateService,
 			{
 				attachTo: document.body,
@@ -35,6 +37,14 @@ describe('CreateService.vue', () => {
 		await router.isReady()
 		clientsStore.clients = [ClientMock]
 		placesStore.places = getPlaces()
+		wpClient.clients = {
+			3103794656: {
+				id: '3103794656',
+				alias: 'Principal',
+				wpNotifications: false,
+				chatBot: false
+			}
+		}
 	})
 	it('an user can show inputs to add service', async () => {
 		await nextTick()
@@ -42,7 +52,7 @@ describe('CreateService.vue', () => {
 		const form = wrapper.findComponent(Form)
 		const input = wrapper.findAll('.form-control')
 		const autoComplete = wrapper.findAllComponents(AutoComplete)
-		expect(field.length).toBe(5)
+		expect(field.length).toBe(6)
 		expect(form.exists()).toBeTruthy()
 		expect(input.length).toBe(4)
 		expect(autoComplete.length).toBe(2)
@@ -61,6 +71,15 @@ describe('CreateService.vue', () => {
 		
 		expect(onSelected).toBeCalledTimes(1)
 	})
+
+	it('calls checkPhoneNoExists when phone number changes', async () => {
+		await nextTick()
+		const checkPhoneNoExists = jest.spyOn(wrapper.vm, 'checkPhoneNoExists')
+		const phoneInput = wrapper.find('input[name="phone"]')
+		await phoneInput.setValue('3101234567')
+		await phoneInput.trigger('input')
+		expect(checkPhoneNoExists).toBeCalledTimes(1)
+	})
 	
 	it('an user can create a new service', async () => {
 		ServiceRepository.create = jest.fn().mockResolvedValue('success')
@@ -72,6 +91,7 @@ describe('CreateService.vue', () => {
 		const comment = 'New comment to service'
 		await wrapper.find('input[name="phone"]').setValue(phone)
 		await wrapper.find('input[name="name"]').setValue(name)
+		await wrapper.find('select[name="wp_client_id"]').setValue('3103794656')
 		const input = wrapper.find('input[name="start_address"]')
 		await input.setValue('mari')
 		await input.trigger('keyup', {
@@ -105,6 +125,7 @@ describe('CreateService.vue', () => {
 		const comment = 'New comment to service'
 		await wrapper.find('input[name="phone"]').setValue(phone)
 		await wrapper.find('input[name="name"]').setValue(name)
+		await wrapper.find('select[name="wp_client_id"]').setValue('3103794656')
 		const input = wrapper.find('input[name="start_address"]')
 		await input.setValue('mari')
 		await wrapper.find('input[name="comment"]').setValue(comment)
@@ -152,6 +173,7 @@ describe('CreateService.vue', () => {
 		const phone = '3100000001'
 		const name = 'Name User'
 		await wrapper.find('input[name="phone"]').setValue(phone)
+		await wrapper.find('select[name="wp_client_id"]').setValue('3103794656')
 		await wrapper.find('input[name="name"]').setValue(name)
 		const input = wrapper.find('input[name="start_address"]')
 		await input.setValue('mari')
@@ -198,7 +220,8 @@ describe('CreateService.vue', () => {
 		ClientRepository.create = jest.fn().mockRejectedValue(new Error('New Error'))
 		const swal = jest.spyOn(Swal, 'fire')
 		await nextTick()
-		
+
+		await wrapper.find('select[name="wp_client_id"]').setValue('3103794656')
 		await wrapper.find('input[name="phone"]').setValue('3100000000')
 		await wrapper.find('input[name="name"]').setValue('Name User')
 		const input = wrapper.find('input[name="start_address"]')